Refactor: eliminate magic numbers and improve precision
- Replace f32 with f64 throughout AppState for consistency (pose ratio, UI values)
- Remove unnecessary f32↔f64 conversions in sidebar UI controls
- Extract magic numbers to named constants in residuals.rs and solver.rs
- Add constants for penalties, tolerances, numerical differentiation, and iteration counts
- Fix cast in plot.rs to eliminate redundant `as f64` conversion
- Improve code readability and maintainability via constant definitions

+++ src/app/plot.rs
@@ -29,7 +29,7 @@ pub fn render_linkage_plot(ui: &mut Ui, state: &mut AppState) {
     };
 
     if should_recompute {
-        let pose = interpolate_pose_for_ratio(&state.config, solution, ratio as f64);
+        let pose = interpolate_pose_for_ratio(&state.config, solution, ratio);
         state.cached_pose_ratio = Some(ratio);
         state.cached_pose = Some(pose);
     }
